+++ stdlib/public/Darwin/simd/Quaternion.swift.gyb
@@ -14,7 +14,6 @@
 
 import Swift
 import Darwin
-import SIMDOperators
 @_exported import simd
 
 %for scalar in ['Float','Double']:
