+++ stdlib/public/Darwin/simd/simd.swift.gyb
@@ -14,7 +14,6 @@
 
 import Swift
 import Darwin
-@_exported import SIMDOperators
 @_exported import simd
 
 public extension SIMD {
