+++ validation-test/Sema/type_checker_perf/fast/rdar46541800b.swift
@@ -1,8 +1,6 @@
 // RUN: %target-typecheck-verify-swift
 // RUN: %target-typecheck-verify-swift -solver-enable-operator-designated-types
 
-import SIMDOperators
-
 func foo(args: [String] = ["A"]) {
   let _ = ["B"]
         + [String](repeating: "C", count: 1 * args.count)
